requirements.txt
dill~=0.3.6
matplotlib~=3.7.1
numpy~=1.24.2
Pillow~=9.4.0
scikit-learn~=1.2.1
tensorboard~=2.12.0
torch~=1.13.1
torchvision~=0.14.1
tqdm~=4.65.0